requirements.txt
alembic==1.5.6
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
fastapi-pagination==0.6.0
FastAPI-SQLAlchemy==0.2.1
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
